SERVER-95804 Fix race in config health observer test (#31217)
This whole test is pretty racy because we mess around with the config
server and hope that the TransientFault turns back into an Ok state
until we disable traffic to all config servers. Most of the time we will
only be able to hit an ActiveFault if all config server nodes are
halting traffic, but it is technically possible for that not to be the
case. When this happens, test assumptions no longer hold.

The goal of this test is to verify that mongos will crash when it cannot
communicate with the config server so this commit modifies the test to
prevent racy conditions when only a subset of the config servers are
down.