SERVER-99589: Consolidate metadata dependency tracking of search and non-search metadata (#31317)
Since Search in SBE is not in use, these changes make it so that Search
in SBE would be less efficient by always building metadata slots. That's
fine since Search in SBE is already far diverged from Search in classic
as used in production. I made a note for posterity if we do ever revisit
Search in SBE.

This patch sets us up for more refactor necessary for proper error
handling in SERVER-99169 and SERVER-93521.