SERVER-99664: Remove coverage on RHEL x86_64 from github_pr_aliases (#31437)
Removing the x86 variant cuts our uploaded tasks to Coveralls by 50%,
which should relieve some pressure and improve throughput.

Code Coverage runs on every PR touch, across Arm and x86, exercising
unittests and handpicked JSTests. This amount of data aggregation
impacts Coveralls' throughput, to the point that data is slow to
resolve, or timeouts occur. We will study the effects of this reduction
(on our data availability and coverage metrics) to see if this temporary
measure should be more permanent (in either direction).