SERVER-97693 Add distinct passthrough test (#31432)
The new jstest was based on other distinct tests in this directory, as
well as jstests/core/query/regex/regex_distinct.js. It extends the
coverage of these tests to include more aggregation cases where
possible, hinting where possible, and allows these tests to run on
sharded & balancing passthroughs (provided that the feature flag is
enabled).