SERVER-95168 Use $rank window function for $rankFusion algorithm (#31262)
The current implementation uses the trick with $group and $push to
compute the rank of documents by adding them all to an array. We believe
it will be better to go with a version that uses $setWindowFields and
$rank to compute the same thing. This has two benefits. First, it won't
be subject to running out of memory if the array gets too big. Second,
it will provide ties with better treatment, giving them equal scores in
the rank fusion algorithm.

It might make performance slightly better, but we haven't seen this in
the data yet. Perhaps in the long term.