SERVER-99029 Histogram CE for AND/OR trees of all sargable ME nodes (#31118)
Given a conjunction of MatchExpressions, function tryHistogramAnd()
attempts to generate and intersect intervals corresponding to groups of
predicates over the same field whenever possible. It then invokes
histogram CE for the resulting simplified intervals. Currently this
function handles only comparison predicates and negation. This commit
extends this function:
* to all nodes for which an interval can be created - this is handled by
function isSargableExpr() in cardinality_estimator.cpp
* and applies the process recursively to nested OR trees under the
conjunction.

Notice that this implementation doesn't fully mimic the functionality of
the index bounds builder - for instance disjunctions of sargable
expressions (or sargable conjunctions) currently will not be folded into
intervals the same way as the index bounds builder.

In order to make the tests that are part of this change work, the commit
also:
* fixes a bug with conjunction breakers not being set for the right type
of QSN nodes,
* fixes a bug in the cost model that didn't account for the cost of the
children of merge sort nodes,
* adjusts several cost coefficients to produce optimal index union plans
with sorting.