SERVER-99742 use removal of file as indicator that resmoke has started in self-tests (#31499)
Using `removeFile` as the indicator to the self-test harness allows the
self-tests to run in parallel without the risk of erroring. Since the
self-tests are run in parallel with resmoke args `--repeatTests=2` and
`--jobs=2`, the same file may be removed/written twice. `removeFile`
will silently pass even if the file does not exist, while `writeFile`
errors if the file already exists.