SERVER-98979 Add sub-level error handling from WT for background compaction (#31544)
This PR is the same as this other
[PR](https://github.com/10gen/mongo/pull/31422#event-16039198218) that
had been reverted [here](https://github.com/10gen/mongo/pull/31494). The
reason for the revert was that in the case where we have an unhandled
`sub_level_err` that is not `WT_NONE`, this would reach a
`MONGO_UNREACHABLE` in `wtRCToStatus`. This case has been changed to
instead return a `Status::OK` to continue error handling as if there
were no sub-level error until this case has been handled.