SERVER-84561 Introduce closeShardCursors() to AsyncResultsMerger/BlockingResultsMerger/DocumentSourceMergeCursors (#31192)
Add method `void closeShardCursors(const stdx::unordered_set<ShardId>&
shardIds);` to `AsyncResultsMerger`, `BlockingResultsMerger` and
`DocumentSourceMergeCursors` .
Closing remote cursors will be useful in change streams contexts.
Currently all change streams that target a single database or collection
are opened on all shards of a sharded cluster. In the future, we want to
restrict the change streams to just the relevant shards. This change is
a step in that direction. We intent to call the new
`closeShardCursors()` method in the future when there are topology
changes in a cluster. Currently the method is not called by any code
outside of tests.

The `closeShardCursors()` method can be used to close and remove all
cursors opened on one of the specified shards. All in-flight requests to
any of these remote cursors will be canceled and their results will be
discarded. All results from the to-be closed remotes that have already
been received by the `AsyncResultsMerger`, but have not been consumed
will be kept. These results can be consumed normally via the
`AsyncResultsMerger::nextReady()` method.
Closing remote cursors is not supported in tailable mode.

In order to be able to remove remote cursors from the
`AsyncResultsMerger`, the `_remotes` container in the
`AsyncResultsMerger` now manages shared pointers to the
`RemoteCursorData` objects. These pointers are now passed around between
the various methods of the `AsyncResultsMerger` instead of the
index/position into the `_remotes` vector. This allows us to remove
remotes from any position inside the vector without having to care about
invalidating indexes/positions into the `_remotes` vector. Instead of
positions into the vector, the private methods of the
`AsyncResultsMerger` pass/receive a shared pointer to the
`RemoteCursorData` in question, which will remain valid even if the
`RemoteCursorData` object was already removed from the `_remotes`
vector.

These pointers are mostly passed around in the `AsyncResultsMerger` by
const reference, and are only copied when sharing ownership is actually
needed. A copy of the shared ptr is made when we capture it in the
callback function that handles receiving a remote's response to the task
executor. This ensures that the remote is still a valid object, even
though it may have been removed from the `_remotes` vector in the
meantime.
We use `boost::intrusive_ptr` and not `std::shared_ptr` to model the
shared ownership in order to keep the costs low, and we only copy the
pointers when necessary. This should keep the costs of using shared
pointers minimal.