SERVER-97373 CBR fallack to multiplanning for unsupported collection and index scans (#31548)
This patch introduces a new error code, `UnsupportedCbrNode`, which is
used by the `CardinalityEstimator` to indicate that plan was unable to
be estimated due to an unsupported node/feature. This allows the query
planner to fallback to multiplanning for that plan. The set nodes banned
is overly conservative, but this approach will allows us to
incrementally and safely implement estimation for them in the future.