SERVER-92115 Improve diagnosability of external_data_source.js (#31542)
This patch is added to improve the diagnosability of the named pipes
test code. All the failures in BF-27419 are related opening the Named
pipes in the C++ test code. This is test only code and is not actually
used in production by ADF. There have been several attempts made in the
past to improve the re-opening logic of named pipes, but to no success.
This patch eliminates the need to re-open named pipes and uses a unique
name for each test run. This will (hopefully) help avoid this problem
all together.

If in case this problem does reproduce after this patch, I've added
additional logs to identify exact operation where the problem might have
occurred on. From the parsley logs, I can see that the Named pipe writer
code is getting hung after logging `The pipe writer thread starts
running`. But it doesn't log
[this](https://github.com/10gen/mongo/blob/0b96c95835955887a37471ad7f1527d12ee4e792/src/mongo/shell/named_pipe_test_helper.cpp#L230)
`pipe opened` statement. I've added additional log line to figure out if
`pipeWriter.open();` is causing the problems or if the cleanup logic in
`NamedPipeOutput` is causing the problems.