SERVER-99526 improve bazel compiledb support (#31525)
This PR completely revamps the bazel compiledb target. It performs all
the work in the python wrapper script, which include:
* generating the compile commands json file
* generate all the generated source
* generate clang tidy config and build checks module (if applicable)

The new compiledb target will by default create a persistent form of
compiledb by leveraging a different output_user_root and different
convienence links. It will always select the --config=dbg configuration
for best results for IDE tools. Because its isolated it will not affect
analysis cache or be affected by any normal build activities.

The wrapper hook also checks for a special arg `--intree_compdb` which
will not use the persistent mode and instead generate the compile
command in an "in tree" fashion, similar to the way it generated before
this PR.