SERVER-98738 Turn off stripping for fastbuild (#31593)
Bazel will by default pass --strip-debug (-S) for fastbuild. We still
want basic symbols for fastbuild and have fission turned on so we don't
want to strip for fastbuild or else gdb can't find the dwo.

You can still pass --linkopt="-s" (strip unneeded) and it works as
expected. Passing --linkopt="-S" (strip debug) gives you an error, so if
you want to turn on debug stripping you should pass --strip=always to
bazel.

This causes mongod linking to go from 550MB to 950MB and for the link
time to go from 10s to 11s in fastbuild.

Also a driveby fix for compiling mozjs for clang under opt for the v5
toolchain.