SERVER-99767: Rename command diagnostics printer to curop diagnostics printer and expose eligibility function (#31451)
This PR is the first step towards adding a new diagnostic printer in
SERVER-93786. It includes the following:`
- The "Command" diagnostic printer is renamed to "CurOp" since it no
longer prints just the command BSON
- Diagnostic logging eligibility checks are factored out to a new
function and exposed for testing and use by future printers
- The new CurOp diagnostic printing files are now part of the "commands"
library, since that is where CurOp is defined and where the CurOp
diagnostic printer is used. These files and associated tests have moved
to the db/ dir.