SERVER-86278 Avoid creating an invalid ReadConcernArgs object with ShardLocal (#31619)
Using a ShardLocal we end up calling the RSLocalClient class which
replaces the opCtx's ReadConcernArgs. However, the existing process
mistakenly may create a ReadConcernArgs in the future which can cause an
invariant to hit since the operation would have both an opTime
specified, and an argsAtClusterTime set, which is an invalid state.

To prevent this we just wait manually and then call the normal wait
which should effectively be a no-op.