SERVER-99690 Fix CBR segfault for queries over clustered collections (#31642)
This patch fixes a segfault for queries which are not express-eligible
but pass 'isIdHackEligibleQuery()' check. For example, find({_id: 1},
{'a.b': 1}). Express does not support dotted projections while idhack
does. The problem was that 'fillOutMainCollectionPlannerParams()' did
not populate 'collStats' when 'isIdHackEligibleQuery()' is true.
However, later on, 'buildIdHackPlan()' fails for this query because the
clustered collection does not have an _id index that idhack can use,
resulting in this query going through the query planner and CBR. The fix
is to populate 'collStats' even if the query passes
'isIdHackEligibleQuery()'.