SERVER-96800 split resharding_coordinator_service.cpp keep large-mem (#31515)
This is a redo of #31269 but kipping
compile_requires_large_memory_sanitizer tag in order to pass
hydrate_bazel_profile_dbg_aubsan builds.

Here we move into a util file all helper content of
resharding_coordinator_service.cpp which is not
part of the ReshardingCoordinator class. This way the large file is
broken down in smaller pieces
which avoids compilation to exceed memory utilization in some cases.