SERVER-99905 Fix object lifetime issues in AsyncResultsMerger (#31637)
Make sure that all instances of `AsyncResultsMerger` are created via a
new factory function, which creates the `AsyncResultsMerger` using
`std::make_shared`.
A copy of the shared_ptr is passed to the lambda function that is used
to handle the asynchronous `getMore` response processing. This makes
sure that the `AsyncResultsMerger` object is still alive and valid when
the asynchronous responses are handled.
Previously, the lambdas could still be unscheduled when the destructor
of the `AsyncResultsMerger` was invoked, which would make the
`AsyncResultsMerger`'s destructor crash with an invariant failure.