SERVER-99150 Add rudimentary support for lock cycle detection in debug builds (#31495)
This adds a rudimentary lock cycle detector exclusively for debug builds
since it is potentially slow and space-hungry.

It detects lock cycles based on preserving lock dependencies across
operations. If we hold a lock A and we're requesting lock B we record
that lock A eventually locks B. At runtime when given two locks we then
ask if the opposite has happened and if not we record the new
dependency.

This detector is rudimentary in the sense that it is space-hungry as it
will keep adding dependencies, and it is in-memory only. However, it has
already proven its worth by detecting numerous issues across the
codebase.

---------

Co-authored-by: Josef Ahmad <josef.ahmad@mongodb.com>