SERVER-99277 Avoid full refresh of sharding metadata on shard removal (#31549)
A hook is triggered on all `mongos` and `mongod` processes whenever a
shard is removed. Previously, this hook caused the entire cached
metadata for databases or collections on the affected shard to be
invalidated, forcing a full cache refresh during the next access. With
the new approach, the cached metadata version is simply advanced,
enabling an incremental refresh instead.