SERVER-98718 Improve startup recovery log in standalone mode (#31704)
I originally intended to only print this when there were any pending
index builds, but realized that it would be good to have this log
message regardless as a way to trace status during startup recovery.
Importantly it is useful to know if any unfinished index builds were
skipped, so this just became a way to log those counts. Of course I'm
also clarifying that unfinished resumable index builds won't be resumed
in standalone mode too.