SERVER-99875 Fix error handling in AsyncResultsMerger (#31621)
Improve error handling in the `AsyncResultsMerger`:
1. Clean up transient errors more consistently. This is achieved by
calling `_cleanUpFailedBatch()` in all locations that can modify the
status of a remote.
2. Move the cleanup from inside `_scheduleGetMores` to the functions
that handle responses.
3. Fix issue when a remote's status was changed between calls to
`ready()` and `nextReady()`, which could lead to tasserts / invariant
failures. This is achieved by setting the global `_status` value to an
error immediately upon any non-transient error, and checking the value
of this `_status` variable inside `ready()` and `nextReady()`.