SERVER-86901 Remove obsolete TSAN suppression (#31777)
Remove a now-obsolete TSAN suppression for
`AsyncResultsMerger::_askForNextBatch(...)`.
That method does not exist since v8.0 anymore.
Furthermore, the AsyncResultsMerger always uses the OperationContext
from the client thread when it calls `scheduleRemoteCommand` on the task
executor. The async tasks that run on the TaskExecutor thread do not
schedule any additional tasks themselves, so they also cannot use a
wrong OperationContext instance.