SERVER-100149 Disable sandboxing on protobuf compiler actions when TSAN is enabled (#31767)
Disable sandboxing on protobuf compiler actions when TSAN is enabled to
avoid an assert from getting hit inside TSAN's initialization when
trying to disable ASLR.

Steps to get down this rabbit hole:
1. The linux execution domain configuration function personality is
returning -1 here inside tsan while it's trying to disable ASLR, likely
because the personality function is not compatible with Bazel's sandbox
2. The proto compiler action runner was ignoring the env variables I set
in our custom compiled proto_compiler when trying to disable TSAN on
protoc
3. I went into the generate_cc grpc source code to find the runner and
threaded in the TSAN_OPTIONS flag to disable bug reporting
4. Unfortunately this wasn't enough since the error we're running into
is an assert, and that assert gets his effectively unconditionally
whenever TSAN is linked into a binary
5. Instead I modified the grpc source code to disable Bazel sandboxing
on the proto compiler action, which fixed the issue

https://jira.mongodb.org/browse/SERVER-100148 is the followup work to
rely on an external Bazel repo instantiation for the proto cpp plugin,
that way we can only enable TSAN on our internal libraries, disabling it
in the proto compiler from the external repository.

Note for reviewers: this sandbox disable ONLY applies to the action to
compile the protobuf into cpp code itself, and does not disable the
sandbox on the logic to compile the resulting cpp code.