SERVER-97605 Kill only the sessions started in the test (#31753)
Kill only the sessions which were started by the test for
internal_transactions_sharded_from_mongod.js
The failures that this test was exhibiting stem from leaving open
sessions
on teardown, and causing a time out for CheckReplDbHash hook which hangs
since
it can't get the global S lock.
This is a retry of #29782, adding a check that the db parameter is
valid.