SERVER-32549 Allow $expr with $in expression to use indexes (#31545)
This PR rewrites expressions of the form `{$expr: {$in: ["$fieldPath",
[...]]}}` into an indexable `MatchExpression` equivalent whenever
possible. It will be rewritten into `{$and: [{value: {$in: [...]}},
{value: {$not: {$type: "array"}}}]}`, where the latter predicate
prevents us from reaching into arrays for comparison to match agg
semantics. The original `$expr` expression will still be
[included](https://github.com/10gen/mongo/blob/master/src/mongo/db/matcher/expression_expr.cpp#L214-L227),
but since we shouldn't need a second filtering level with the
constraints, it can be removed in a future ticket.

The requirements are:
- `lhs` is a field path
- `rhs` is a constant array

If the following types are present in the `rhs` array, the expression is
ineligible for rewrites and will be unchanged. The query will still use
a collection scan.
- Regex
- Null/Undefined
- Array