SERVER-90992 Detect collections that rely on snapshot stability being used after abandoning the snapshot (#31202)
This introduces the notion of ConsistentCollection which is a mechanism
that detects collection references being leftover once we abandon a
snapshot.

This is useful for lock-free reads since they may cause a Collection
instance to be created and tied to the lifetime of the current snapshot
in the event of concurrent DDL operations.

The changes introduced here are the minimal set of required changes to
avoid stale references. Note that using the old mechanism of manually
constructing a CollectionPtr is still valid. Additionally, this change
should not have a noticeable delay in release builds since the reference
counting is only present in debug builds.