SERVER-71323 Change FLEClientCrypto::transformPlaceholders to use libmongocrypt for encryption (#31742)
libmongocrypt's encryption pipeline works differently in a few subtle
ways to mongo's. The most important of these is that libmongocrypt has
much stricter criteria for when we should include
`encryptionInformation` on the final payload, meaning we have to change
a handful of tests which relied on the old behavior.

libmongocrypt is pulled from master in order to get text search query
type definitions.

---------

Co-authored-by: Erwin Pe <erwin.pe@mongodb.com>