SERVER-100305 Fix gRPC warnings and errors when compiling with the v5 toolchain (#31881)
gRPC fails to compile right now on the v5 toolchain and throws a bunch
of other warnings. Disable all of the errors that are failing the build
and cleanup as much of the warnings as possible.

This PR also greatly simplifies our current handling of third party
warnings, moving them to a consolidated location in the toolchain config
that conditionally enables each set based on the compiler that the
toolchain is being configured for.

This has the downside of not configuring warnings for the specific third
party library itself, but this is a great tradeoff for the added
simplicity in implementation and making it easier to audit for more
dangerous warnings. We primarily care about the nature of the warning
being ignored, not which library the warning is in. For example, we
don't care if a -Wdeprecated is triggered in tcmalloc or zlib. What we
care about is the severity of the warning type being ignored. If we
intend to go through and audit third party libraries for warnings, we'd
likely do a scan per ignored instance focusing on the most dangerous
warnings.