SERVER-99675: Implement typo suggestion for misspelled $rankFusion weights (#31931)
[SERVER-99675](https://jira.mongodb.org/browse/SERVER-99675)

Builds on top of the previous https://github.com/10gen/mongo/pull/31478
that allows for a subset of weights to be specified in the
combination.weights object. Now, instead of directly rejecting the query
once a non-existent pipeline is found, all the specified weights are
processed to provide better feedback to the user.

The error message thats provided to the user attempts to provide the
best possible suggestion (a valid unmatched pipeline) for each invalid
weight specified.

A test is provided that validates that the right suggestions are
computed for different provided weight combinations.

An example error message for a $rankFusion pipeline like:
```
{
    $rankFusion: {
         input: {
             pipelines: {
                 vector: [...],
                 search: [...]
            }
        },
        combination: {
            weights: {
                 vecor: 0.5,
                 searchh: 0.3
            }
        }
    }
}
```
would produce error message:
```
$rankFusion stage contained (2) weight(s) in 'combination.weights' that did not reference valid pipeline names. Suggestions for valid pipeline names: {provided: 'vecor' -> suggested: 'vector'}, {provided: 'searchh' -> suggested: 'search'}.
```
In cases where a single weight could have multiple equally good
suggestions, $rankFusion will produce an error structured like:
```
$rankFusion stage contained (1) weight(s) in 'combination.weights' that did not reference valid pipeline names. Suggestions for valid pipeline names: (provided: 'fite' -> suggestions: [kite, mite, nite])
```
This implementation uses a preexisting algorithm called the "Levenshtein
Distance" to numerically score how similar two strings are (where
smaller scores are better, and two equal strings score 0). The
"Levenshtein Distance" computes the minimum number of edits needed to
transform one string into another, so that we can compare which of a
possible set of strings (the valid pipeline names in this case) is most
similar to a single reference string (the invalid weight in this case).
Here is a helpful blogpost that explains how the algorithm works (in
addition to my comment):
https://blog.paperspace.com/measuring-text-similarity-using-levenshtein-distance/

Also, I used [Leetcode #72 Edit
Distance](https://leetcode.com/problems/edit-distance/description/) to
test my levenshteinDistance function. I suppose if we wanted to be more
resilient to future changes, we could unit test this function
independently with some string pair test cases.