SERVER-98654 Allow all WT operations to be interrupted (#31127)
Unconditionally adds an interruptible to the WT_SESSION to eventually
allow more WT events to be interrupted. Currently doesn't modify the
callback `mdb_handle_general` so this will continue to only interrupt
compact, but once WT implements more general events we can choose
whether those are interrupted as well.

As no new events are implemented the existing testing for compact is all
that can be done. Once possible additional testing to be added in
[SERVER-99264](https://jira.mongodb.org/browse/SERVER-99264)