SERVER-99957 Fix test-only data races in replication_coordinator_impl_test (#31829)
The v5 toolchain reports some data races in this test due to cases where
we write to a variable called `deadline` in the main thread, start
another thread capturing `deadline` by reference, and then write to
`deadline` again in the main thread. There was also one case where the
started thread is the one that writes to deadline and the main thread is
the one reading from it.

There are lots of other cases similar to this one in the test that
aren't racy - in these cases, thread starts and joins provide enough
synchronization. This change doesn't touch those cases.