SERVER-99502 Fix static initialization order fiasco in query_analysis_sampler_test (#31824)
The v5 toolchain reports a static initialization order fiasco on the
initialization of `smoothingFactor` because it loads from a global
server parameter that may not be initialized when the load happens. It's
only used in one function and shouldn't be getting changed dynamically
during the test, so it's an easy fix by moving the initialization of
`smoothingFactor` into the function that uses it.