SERVER-99617 Boost DISTINCT_SCAN productivity score (#31618)
This effectively doubles the productivity score of a DISTINCT_SCAN in an
aggregation multiplanning context.

**Motivation**: unlike in a distinct command, where we always try to
dedup the output of a distinct scan & an index scan (which means that
the productivity of a plan is a good proxy for query efficiency), in an
aggregation context, we perform a different operation on the output of a
cursor depending on whether or not a DISTINCT_SCAN was used (i.e.
$groupByDistinct vs $group). A $groupByDistinct rewrite is substantially
more performant than a regular $group, even in cases where an index scan
is twice as productive.

Performance tests show an index scan plan with 2x the productivity score
as a distinct scan can also be twice as slow, even when there are no
duplicates. However, performance tests also show that for highly
selective plans, an index scan can be 80% faster than an equivalent
distinct scan. For this reason, we still want to trial index scans,
while heavily favoring distinct scans.