SERVER-95382 Introduce python module to find the mongo toolchain (#31229)
This change introduces a new module in `buildscripts/mongo_toolchain.py`
that can be used to find a mongo toolchain in several common execution
environments. For running standalone scripts, it can find a toolchain
under /opt/mongodbtoolchain or it can use one downloaded by bazel. For
v4, /opt is the default and for v5, downloading from bazel is the
default. For scripts running under bazel, bazel automatically downloads
the toolchain, puts in the scripts runfiles, and sets an environment
variable that the module uses the find it. This requires using a new
`mongo_py_cxx_test` rule for such tests.

A shell script could also use mongo_toolchain.py to find a toolchain.
This change doesn't do that for any shell scripts or introduce the bazel
macros necessary for them to work with it under bazel.