SERVER-100395 Return a parse error when objects are nested too deeply in JParse (#32026)
When descending into objects or arrays, JParse makes a recursive call,
leading to a risk of stack exhaustion when parsing deeply-nested JSON
objects or arrays. This change imposes a fixed arbitrary limit on the
number of nested entities we support. Any nesting deeper than that is
treated as a parsing error.

This is done by passing a `depth` argument down through the recursive
calls. When calling `value`, we don't increment the depth, but when
calling any other function that needs a depth parameter, we do. This is
because each actual entity involves a call to both `value` and `object`
and `array`, so incrementing in both places would double count the
degree of nesting.

Besides `array` and `object`, `dbRef` and `dbRefObject` also take a
depth parameter as they call into `value` to parse the `$id` field and
therefore are at risk of stack exhaustion. Objects having deeply-nested
objects as an `$id` would presumably be semantically invalid, but the
parser happily parses them, so is still vulnerable in that situation.