SERVER-100313 Move PlanCacheCallbacksImpl to query exec (#31896)
For [SERVER-87128](https://jira.mongodb.org/browse/SERVER-87128) we need
to notify waiting threads that a plan cache key has been added to the
plan cache. The best way to do it is to leverage PlanCacheCallbacks.

Unfortunately, PlanCacheCallbacksImpl is implemented in query_plan_cache
library while it is used in query_exec. To notify all waiting threads I
need to add a dependency defined in query_exec. It seems that the best
place for the PlanCacheCallbacksImpl is query_exec, where it is used,
and its interface PlanCacheCallbacks should remain in query_plan_cache.

This purely refactoring change would allow to streamline implementation
of [SERVER-87128](https://jira.mongodb.org/browse/SERVER-87128) and
simplify its code review.