SERVER-100109 Disable ASAN option detect_stack_use_after_return for v5 toolchain (#31925)
In clang 15, detect_stack_use_after_return became the default on linux.
One effect from this option is that automatic storage duration variables
(colloquially, "variables on the stack") are not actually allocated on
the stack. This breaks sigaltstack_location_test, which uses an
automatic storage duration variable to find where the stack is. That's a
test-only issue. It also breaks StackLocator, which uses a similar
technique to estimate how much space remains on the stack. This is used
in MozJS to set limits on the javascript stack size.

In the v4 toolchain this change is a no-op since that option was already
disabled there by default. For v5, the option disables the new check. We
should aim to turn it back on once the issues can be fixed.