SERVER-99596: Refactor metadata field dependency validation (#31488)
This is a significant refactor of how we validate references to metadata
fields (via the $meta operator under $project or $sort). The changes are
a bit dubious, so I've left a detailed explanation. Sorry it's long, but
I think (hope!) this will help review.

### Previous State
The current codebase has comments like this attached to many uses of
`DepsTrackers`:
```
// We only need to know what variable dependencies exist in the subpipeline. So without
// knowledge of what metadata is in fact unavailable, we "lie" and say that all metadata
// is available to avoid tripping any assertions.
```
We initialize the tracker like `DepsTracker(/*unavailableMetadata=*/
DepsTracker::kNoMetadata)` to say "since _no_ metadata is _unavailable_,
any metadata references are valid. This is a "hack" to circumvent
metadata field validation in cases when we don't want to do such
validating. This is confusing for many reasons, including the double
negative that _no_ metadata is _unavailable_.

### New State
First of all, we remove the double negative so that we simply track
which metadata _is_ explicitly available to the pipeline.
Then, we track a const set of fields that should be validated
(validation opt-in: right now just "textScore", "geoNearDistance", and
"geoNearPoint"; we're expanding that list in followup tickets
SERVER-99169 and SERVER-93521).

Lastly, we force uses of `DepsTracker` to more clearly opt in or out of
metadata field validation. If the constructor parameter
`availableMetadata` is given a `QueryMetadataBitSet`, then we _do_
perform validation; if it's `DepsTracker::NoMetadataValidation`, we skip
the validation (since we don't need to track which fields are available
or not). boost::none is now the default since most uses of the
`DepsTracker` were actually skipping the validation.

What was `DepsTracker(DepsTracker::kNoMetadata)` for the purposes of
skipping validation is now `DepsTracker deps;` (default constructor) in
most cases. What was `DepsTracker(DepsTracker::kAllMetadata)` for the
sake of saying "all metadata is unavailable" is now
`DepsTracker(DepsTracker::kNoMetadata` for the translation "no metadata
is available". etc.

### Are we sure it's safe to invert unavailableMetadata ->
availableMetadata?
The "double negative" of it all was introduced in SERVER-12335. I
couldn't find the rationale for why, but I _think_ it's so that we have
to "opt in" to fields that should be validated by marking them
explicitly unavailable; then all other fields are default available. We
get around this in this patch by tracking the set of fields that should
be validated, so it's fine if other fields aren't marked explicitly
available.

In advance of this ticket, we added a property-based test in
SERVER-100045 to thoroughly test the behavior of $meta validation; we
should feel confident about this change with no failures introduced in
that test.