SERVER-96479 Account for storage metrics when recovery unit is swapped during spilling and capped deletes (#32086)
We use a separate recovery unit for capped deletes and query spilling.
We should stash the storage metrics stored on the recovery unit when
swapping out the recovery unit and unstash these metrics when restoring
the recovery unit. This change ensures that curop reports the cumulative
metrics that have accrued during the operation even if it was on a
different recovery unit.