SERVER-100630 Reset sharded collection on transient transaction error in mongos_api_params_util.js (#32196)
When we hit a transient transaction error, the test drops the database
and attempts to retry the operation. However if the test case requires a
sharded collection and asserts on the primary shard's logs, then
dropping the database also means that we are no longer guaranteed to run
against the original shard primary.