SERVER-97941 Add strEncodeVersion field to EncryptedFieldConfig (#32149)
Libmongocrypt changes: https://github.com/mongodb/libmongocrypt/pull/949

This field is optional, and not used for anything by the server. It is
simply passed through on the EncryptedFieldConfig. We add a test-only
check in client-side encryption that after all processing, it is true
that this field is always sent when we have an field with a text search
query type.

Note: This is based on the linked PR's version of libmongocrypt, which
is not yet merged. Once it is merged, I will pin that commit.

---------

Co-authored-by: Erwin Pe <erwin.pe@mongodb.com>