SERVER-98411: Avoid a use-after-free memory bug when the StorageEngineImpl constructor throws. (#32218)
When `StorageEngineImpl` constructor throws, the `OperationContext`
instance passed to the constructor of `StorageEngineImpl` tries to
delete the `WiredTigerSession` inside it after the
`WiredTigerConnection` instance has already been deleted. Fixing it by
using a ScopeGuard.

I explicitly threw an exception in the `StorageEngineImpl` constructor
and started mongod in address sanitizer mode to reproduce the issue and
also to verify that the fix works.