SERVER-94718 Create a property-based test to assert correct results when hinting different indexes (#32154)
Some types of query plans are rarely selected and are undertested
end-to-end. This is a PBT hints every index available to trigger as many
plans as we can and assert they produce the correct output.

In the future, we'll have more robust hinting that can force
intersection or union plans. When that feature is available this PBT can
truly hint every plan and will provide more coverage. I've added a TODO
for this.

This discovered SERVER-99889.