SERVER-98571 Fix scale parameter for dbStats (#31830)
Previous change was autoreverted, some test environments don't seem to
work well when toeing the line of the maximum safe integer. I chose a
max scale of one petabyte which steps back a couple bits and is still
far larger than is likely to ever be needed. Also, added a backport
required for multiversion tests.