SERVER-100820 Fix many validate_mongocryptd invocation bugs (#32367)
There are at least 4 separate bugs with this script that's prevented it
from being run in at least several years. Primarily the fact that it's
checking -eq instead of -ge given that we will always haven an even
number of mongocryptd (either 0, 2, or 4) will prevent it from ever
being a no-op. Somehow the release pipeline is getting a single
mongocryptd which is making it error out later (no active venv,
evergreen yaml is no longer trivially parsable, etc)