SERVER-100619 Introduce jsCore passthrough for CBR with histogram CE (#32339)
This PR achieves the following:

* Introduce an override script which intercepts CRUD commands and runs
the analyze command over all indexed fields, then dispatches the
original command. This script can be reused in subsequent passthroughs
which want to get coverage for histogram based cardinality estimation.
* Integration test that verifies the behavior of the override script.
* Introduces jsCore passthrough which runs with `automaticCE`.
Currently, this is necessary because the semantics of `histogramCE` mode
will assert if encountering a non-sargable predicate. This will use
histograms where possible and fallback to heuristics.
* Exclude failing tests from the suite and categorize them by
ticket/reason.
* Add the passthrough to the query patch variants and schedule them to
run nightly.